+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-web/nifi-web-api/src/main/java/org/apache/nifi/audit/ProcessorAuditor.java
@@ -170,6 +170,7 @@ public class ProcessorAuditor extends NiFiAuditor {
                 if (operation != null) {
                     // clear the value if this property is sensitive
                     final PropertyDescriptor propertyDescriptor = processor.getPropertyDescriptor(property);
+                    // Evaluate both Property Descriptor status and whether the client requested a new Sensitive Dynamic Property
                     if (propertyDescriptor != null && (propertyDescriptor.isSensitive() || sensitiveDynamicPropertyNames.contains(property))) {
                         if (newValue != null) {
                             newValue = SENSITIVE_VALUE_PLACEHOLDER;
